chore: break out custom node components
